+++ .github/CHANGELOG.md
@@ -235,6 +235,7 @@ Always start an item with "1."; GitHub will number the rendered list automatical
 1. [EFB] Add FAA and LIDO charts supplied by MSFS2024 - @tracernz (Mike)
 1. [A32NX/FLIGHT MODEL] Updated A32NX flight model for MSFS 2024 - @donstim (donbikes)
 1. [A32NX/FWS] Add `BARO VALUE DISAGREE` master caution, rename `ALTI DISCREPANCY` master caution to `ALT DISCREPANCY` - @FozzieHi (fozzie)
+1. [A32NX/FWS] Add intermediate callouts to FWC - @BravoMike99 (bruno_pt99)
 
 ## 0.14.0